Responding in earnest to https://twitter.com/twolivesleft/status/1328178489258692608
Its kind of cool that a tweet can lead to change, so here you go, the removal of the ugly red bang emoji.

But, here's a bit of the back story.

Its true that I dislike negative logic. I believe we should strive for code to be able to read left to right and sound like English.
The tweet responders reacted so dramatically to the red bang, but its not like it was forced on people, you could still use the regular one.
The red bang was introduced into my code to call attention to negative code as a reminder to change it.
Many things were made clearer as a result, but really I left it hanging for a long time and now I am also glad to have it gone.